fix: refine selector picker states
