+++ apps/web/components/task-create-dialog-remote-repo-chip.tsx
@@ -10,12 +10,12 @@ import { Button } from "@kandev/ui/button";
 import { Popover, PopoverContent, PopoverTrigger } from "@kandev/ui/popover";
 import { Spinner } from "@kandev/ui/spinner";
 import { Tooltip, TooltipContent, TooltipTrigger } from "@kandev/ui/tooltip";
+import { Pill } from "@/components/task-create-dialog-pill";
 import {
-  Pill,
   branchToOption,
   computeBranchPlaceholder,
   sortBranches,
-} from "@/components/task-create-dialog-pill";
+} from "@/components/task-create-dialog-branch-options";
 import { scoreBranch } from "@/lib/utils/branch-filter";
 import type {
   RemoteRepository,
