+++ apps/web/e2e/tests/chat/mobile-model-selector.spec.ts
@@ -121,18 +121,21 @@ test.describe("Mobile chat model selector", () => {
     await expect(trigger).toHaveText("Mock Fast", { timeout: 15_000 });
 
     await trigger.tap();
-    await expect(testPage.getByRole("option", { name: /Mock Smart/ })).toBeVisible({
+    const modelListbox = testPage.getByRole("listbox");
+    await expect(modelListbox.getByRole("option", { name: /Mock Smart/ })).toBeVisible({
       timeout: 5_000,
     });
-    const modelOptions = testPage.getByRole("option");
+    const modelOptions = modelListbox.getByRole("option");
     await expect(modelOptions.first()).toContainText("Mock Fast");
     await expect(modelOptions.first()).toHaveClass(/bg-card/);
     await expect(modelOptions.first()).toHaveClass(/border-primary\/50/);
     await expect(
-      testPage.getByRole("option", { name: /Mock Fast/ }).getByTitle("Fast mock model for testing"),
+      modelListbox
+        .getByRole("option", { name: /Mock Fast/ })
+        .getByTitle("Fast mock model for testing"),
     ).toBeVisible();
     await expect(
-      testPage
+      modelListbox
         .getByRole("option", { name: /Mock Smart/ })
         .getByTitle("Smart mock model for testing"),
     ).toBeVisible();
