+++ docs/plans/selected-option-picker-prominence/task-01-selected-option-picker-ui.md
@@ -39,6 +39,8 @@ cd apps/web && rtk pnpm run typecheck
 - `apps/web/components/combobox.test.tsx`
 - `apps/web/components/task-create-dialog-pill.tsx`
 - `apps/web/components/task-create-dialog-pill.test.tsx`
+- `apps/web/components/task-create-dialog-branch-options.tsx`
+- `apps/web/hooks/use-pill-tooltip-suppression.ts`
 - `apps/web/components/model-config-selector.tsx`
 - `apps/web/components/model-config-selector.test.tsx`
 - `apps/web/components/task/mode-selector.tsx`
@@ -69,9 +71,13 @@ synchronized task/plan status.
 
 ## Results
 
-Focused implementation tests pass: 6 files, 51 tests. TypeScript typecheck
+Focused implementation tests pass: 6 files, 55 tests. TypeScript typecheck
 passes. The shared combobox, compact task-create pills, model/config picker,
 mode menu, and dedicated branch picker now share selected-first ordering and
-persistent selected-row styling. Mobile behavior keeps the existing command
-and branch-sheet compositions, with 44px touch rows and no new overflow. The
-changed frontend files pass targeted ESLint with zero warnings and Prettier.
+persistent selected-row styling. Keyboard navigation coverage confirms cmdk's
+transient highlight remains distinct from the persistent current-row surface.
+The model content, branch option utilities, and pill tooltip suppression are
+split into focused modules so the changed files remain within repository size
+limits. Mobile behavior keeps the existing command and branch-sheet
+compositions, with 44px touch rows and no new overflow. The changed frontend
+files pass targeted ESLint with zero warnings and Prettier.
